requirements.txt
torch
transformers
gradio
mlflow>=3.5.1
openai>=1.17
matplotlib
numpy
pandas
boto3
psycopg2-binary
python-dotenv
psutil
requests
